move get_priority
