+++ kong/router/transform.lua
@@ -3,6 +3,7 @@ local buffer = require("string.buffer")
 local lrucache = require("resty.lrucache")
 local ipmatcher = require("resty.ipmatcher")
 local utils = require("kong.router.utils")
+local tb_nkeys = require("table.nkeys")
 
 
 local type = type
@@ -11,7 +12,7 @@ local tonumber = tonumber
 local ipairs = ipairs
 local fmt = string.format
 local byte = string.byte
-local band, lshift, rshift = bit.band, bit.lshift, bit.rshift
+local bor, band, lshift, rshift = bit.bor, bit.band, bit.lshift, bit.rshift
 
 
 local is_regex_magic  = utils.is_regex_magic
@@ -421,6 +422,196 @@ do
 end
 
 
+local stream_get_priority
+do
+  -- compatible with http priority
+  local STREAM_SNI_BIT = lshift_uint64(0x01ULL, 61)
+
+  -- IP > PORT > CIDR
+  local IP_BIT         = lshift_uint64(0x01ULL, 3)
+  local PORT_BIT       = lshift_uint64(0x01ULL, 2)
+  local CIDR_BIT       = lshift_uint64(0x01ULL, 0)
+
+  local function calc_ip_weight(ips)
+    local weight = 0x0ULL
+
+    if is_empty_field(ips) then
+      return weight
+    end
+
+    for i = 1, #ips do
+      local ip   = ips[i].ip
+      local port = ips[i].port
+
+      if ip then
+        if ip:find("/", 1, true) then
+          weight = bor(weight, CIDR_BIT)
+
+        else
+          weight = bor(weight, IP_BIT)
+        end
+      end
+
+      if port then
+        weight = bor(weight, PORT_BIT)
+      end
+    end
+
+    return weight
+  end
+
+  stream_get_priority = function(snis, srcs, dsts)
+    local match_weight = 0x0ULL
+
+    -- [sni] has higher priority than [src] or [dst]
+    if not is_empty_field(snis) then
+      match_weight = STREAM_SNI_BIT
+    end
+
+    -- [src] + [dst] has higher priority than [sni]
+    if not is_empty_field(srcs) and
+       not is_empty_field(dsts)
+    then
+      match_weight = STREAM_SNI_BIT
+    end
+
+    local src_bits = calc_ip_weight(srcs)
+    local dst_bits = calc_ip_weight(dsts)
+
+    local priority = bor(match_weight,
+                         lshift_uint64(src_bits, 4),
+                         dst_bits)
+
+    return priority
+  end
+end
+
+
+local MAX_HEADER_COUNT = 255
+
+
+local PLAIN_HOST_ONLY_BIT = lshift_uint64(0x01ULL, 60)
+local REGEX_URL_BIT       = lshift_uint64(0x01ULL, 51)
+
+
+-- convert a route to a priority value for use in the ATC router
+-- priority must be a 64-bit non negative integer
+-- format (big endian):
+--  0                   1                   2                   3
+--  0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
+-- +-----+-+---------------+-+-------------------------------------+
+-- | W   |P| Header        |R|  Regex                              |
+-- | G   |L|               |G|  Priority                           |
+-- | T   |N| Count         |X|                                     |
+-- +-----+-+-----------------+-------------------------------------+
+-- |  Regex Priority         |   Max Length                        |
+-- |  (cont)                 |                                     |
+-- |                         |                                     |
+-- +-------------------------+-------------------------------------+
+local function get_priority(route)
+  local snis = route.snis
+  local srcs = route.sources
+  local dsts = route.destinations
+
+  -- stream expression
+
+  if not is_empty_field(srcs) or
+     not is_empty_field(dsts)
+  then
+    return stream_get_priority(snis, srcs, dsts)
+  end
+
+  -- http expression
+
+  local methods = route.methods
+  local hosts   = route.hosts
+  local paths   = route.paths
+  local headers = route.headers
+
+  local match_weight = 0  -- 0x0ULL
+
+  if not is_empty_field(methods) then
+    match_weight = match_weight + 1
+  end
+
+  if not is_empty_field(hosts) then
+    match_weight = match_weight + 1
+  end
+
+  local headers_count = is_empty_field(headers) and 0 or tb_nkeys(headers)
+
+  if headers_count > 0 then
+    match_weight = match_weight + 1
+
+    if headers_count > MAX_HEADER_COUNT then
+      ngx.log(ngx.WARN, "too many headers in route ", route.id,
+                        " headers count capped at ", MAX_HEADER_COUNT,
+                        " when sorting")
+      headers_count = MAX_HEADER_COUNT
+    end
+  end
+
+  if not is_empty_field(snis) then
+    match_weight = match_weight + 1
+  end
+
+  local plain_host_only = type(hosts) == "table"
+
+  if plain_host_only then
+    for _, h in ipairs(hosts) do
+      if h:find("*", nil, true) then
+        plain_host_only = false
+        break
+      end
+    end
+  end
+
+  local uri_length = 0
+  local regex_url = false
+
+  if not is_empty_field(paths) then
+    match_weight = match_weight + 1
+
+    local p = paths[1]
+
+    if is_regex_magic(p) then
+      regex_url = true
+
+    else
+      uri_length = #p
+    end
+
+    for i = 2, #paths do
+      p = paths[i]
+
+      if regex_url then
+        assert(is_regex_magic(p),
+               "cannot mix regex and non-regex paths in get_priority()")
+
+      else
+        assert(#p == uri_length,
+               "cannot mix different length prefixes in get_priority()")
+      end
+    end
+  end
+
+  local match_weight   = lshift_uint64(match_weight, 61)
+  local headers_count  = lshift_uint64(headers_count, 52)
+
+  local regex_priority = lshift_uint64(regex_url and route.regex_priority or 0, 19)
+  local max_length     = band(uri_length, 0x7FFFF)
+
+  local priority = bor(match_weight,
+                       plain_host_only and PLAIN_HOST_ONLY_BIT or 0,
+                       regex_url and REGEX_URL_BIT or 0,
+                       headers_count,
+                       regex_priority,
+                       max_length)
+
+  return priority
+end
+
+
 return {
   LOGICAL_OR  = LOGICAL_OR,
   LOGICAL_AND = LOGICAL_AND,
@@ -432,6 +623,7 @@ return {
   gen_for_field = gen_for_field,
 
   get_expression = get_expression,
+  get_priority = get_priority,
 
   lshift_uint64 = lshift_uint64,
 }
